Enable audio offloading with video
According to our system_props, we're trying to make use of audio
offloading with streaming video. The prop for this is:
av.streaming.offload.enable, which is set to true in our case.
This doesn't actually do anything as it depends on audio.offload.video,
which should be enabled (defaults to false if it isn't set).

If we want to make use of this feature, this needs to be merged.
If not, then we maybe should remove the prop from our system_props.mk

Change-Id: Iec41f30d6e3cebdd52facbcd3483ba90e15b29d2

Conflicts:
	system_prop.mk